REFACTOR: Gate serialization and processor configuration
