+++ src/qib/backend/__init__.py
@@ -1,5 +1,6 @@
 from qib.backend.options import Options
+from qib.backend.processor_configuration import ProcessorConfiguration, GateProperties
 from qib.backend.experiment import Experiment, ExperimentResults, ExperimentStatus, ExperimentType
-from qib.backend.quantum_processor import QuantumProcessor, ProcessorConfiguration, GateProperties
+from qib.backend.quantum_processor import QuantumProcessor
 from qib.backend.tensor_network_processor import TensorNetworkProcessor
 from qib.backend import wmi, qiskit
