+++ src/qib/util/const.py
@@ -11,7 +11,7 @@ GATE_CY: str = 'cy'
 GATE_CZ: str = 'cz'
 GATE_H: str = 'h'
 GATE_ID: str = 'id'
-GATE_R: str = 'u'
+GATE_R: str = 'u3'
 GATE_RX: str = 'rx'
 GATE_RY: str = 'ry'
 GATE_RZ: str = 'rz'
